+++ src/components/faq/index.jsx
@@ -10,7 +10,7 @@ function Faq({ faqs }) {
   };
 
   return (
-    <div className="max-w-[1160px] mx-auto px-4 py-8">
+    <div className="max-w-[1160px] mx-auto">
       <div className="grid grid-cols-1 md:grid-cols-2 gap-4 place-items-start">
         {faqs.map((faq, index) => (
           <div
